Remove a bunch of old code
Some was moved to larpix-scripts, and some just turned out to not be
useful.
